Check for overflows with the RLE row lengths
If the image data does not compress the row byte length could exceed
short.MaxValue.

+++ src/RLEHelper.cs
@@ -141,14 +141,14 @@ namespace AbrFileTypePlugin
 
         ////////////////////////////////////////////////////////////////////////
 
-        public static int EncodedRow(Stream stream, ReadOnlySpan<byte> row)
+        public static long EncodedRow(Stream stream, ReadOnlySpan<byte> row)
         {
             long startPosition = stream.Position;
 
             RlePacketStateMachine machine = new(stream);
             machine.PushRow(row);
 
-            return (int)(stream.Position - startPosition);
+            return stream.Position - startPosition;
         }
 
         ////////////////////////////////////////////////////////////////////////
